+++ constant.js
@@ -1,0 +1,9 @@
+const SALT_KEY = 'salt_key'
+const IV_KEY = 'iv_key'
+const PASSWORD_LENGTH = 5
+
+module.exports = {
+  SALT_KEY,
+  IV_KEY,
+  PASSWORD_LENGTH
+}
