+++ lib/index.js
@@ -1,0 +1,8 @@
+'use strict'
+const utils = require('./utils')
+const crypto = require('./crypto')
+
+module.exports = {
+  crypto,
+  utils
+}
